fix: S24.12 sign-conversion on ~BIT after enum + format reflow
CI surfaced two regressions from the previous slices that the local
gcc Tier-A checks didn't catch:

* build-freertos-host-tdd-plustcp — the CmsdkUart register-macro -> enum
  conversion lost the `U` suffix on the bit values (C enum constants
  are `int` per C99 6.7.2.2). The CMSDK fake's three `fake.state &= ~BIT`
  expressions now produce an `int` value (e.g. `~1` = `-2`) that
  -Werror=sign-conversion catches on the uint32_t assignment. Fixed by
  casting the operand to (uint32_t) at the three sites. The enum form
  stays (the SoC-manual grep traceability + type-check the conversion
  bought is preserved); only the bit-clearing expressions are explicit
  about the unsigned mask. Verified locally with the freertos-host
  container.

* analyze-format — bulk inline-NOLINT deletions left double-blank lines
  where a standalone NOLINTBEGIN / NOLINTEND sat alone on its own
  line. clang-format collapses consecutive blank lines to one. Reflowed
  37 Tests/* files via `clang-format -i`.

Process learning: per-commit Tier-A (gcc-only) wasn't enough to catch
the CmsdkUart sign-conversion. The freertos-host image is reachable
locally via the same docker compose recipe CI uses
(`.devcontainer/docker-compose.yml` `freertos-host` service); running
both gcc and freertos-host before pushing is the right pre-push step
for any change that touches `Bdd/Targets/FreeRtos/Common/` or
`Tests/FreeRtos/`.

+++ Tests/FreeRtos/CmsdkUartFake.c
@@ -47,7 +47,7 @@ static uint32_t Fake_Read32(uintptr_t address)
         }
         if (fake.readsRemainingBeforeTxReady == 0)
         {
-            fake.state &= ~TX_FULL_BIT;
+            fake.state &= ~(uint32_t) TX_FULL_BIT;
         }
         if (fake.readsRemainingBeforeRxReady > 0)
         {
@@ -64,7 +64,7 @@ static uint32_t Fake_Read32(uintptr_t address)
         if ((fake.state & RX_FULL_BIT) != 0U)
         {
             result = fake.receivedByte;
-            fake.state &= ~RX_FULL_BIT;
+            fake.state &= ~(uint32_t) RX_FULL_BIT;
         }
     }
     else if (offset == CTRL_OFFSET)
@@ -181,7 +181,7 @@ void CmsdkUartFake_SetReceivedByte(char byte)
     /* Clear leftover RX-ready state from a prior arm so back-to-back calls
      * don't carry RX_FULL_BIT or a stale countdown into the new mode. */
     fake.receivedByte = (uint32_t) (unsigned char) byte;
-    fake.state &= ~RX_FULL_BIT;
+    fake.state &= ~(uint32_t) RX_FULL_BIT;
     fake.readsRemainingBeforeRxReady = 0;
     if (fake.readsBeforeRxReadyDefault == 0)
     {
